--- a/src/accelerate/test_utils/testing.py
+++ b/src/accelerate/test_utils/testing.py
@@ -40,6 +40,7 @@ from ..utils import (
     is_datasets_available,
     is_deepspeed_available,
     is_dvclive_available,
+    is_import_timer_available,
     is_mlu_available,
     is_mps_available,
     is_npu_available,
@@ -52,7 +53,6 @@ from ..utils import (
     is_torch_xla_available,
     is_torchvision_available,
     is_transformers_available,
-    is_tuna_interpreter_available,
     is_wandb_available,
     is_xpu_available,
     str_to_bool,
@@ -378,12 +378,12 @@ def require_pippy(test_case):
     return unittest.skipUnless(is_pippy_available(), "test requires pippy")(test_case)
 
 
-def require_tuna_interpreter(test_case):
+def require_import_timer(test_case):
     """
     Decorator marking a test that requires tuna interpreter installed. These tests are skipped when tuna isn't
     installed
     """
-    return unittest.skipUnless(is_tuna_interpreter_available(), "test requires tuna interpreter")(test_case)
+    return unittest.skipUnless(is_import_timer_available(), "test requires tuna interpreter")(test_case)
 
 
 _atleast_one_tracker_available = (