--- a/src/accelerate/utils/imports.py
+++ b/src/accelerate/utils/imports.py
@@ -81,8 +81,8 @@ def get_ccl_version():
     return importlib.metadata.version("oneccl_bind_pt")
 
 
-def is_tuna_interpreter_available():
-    return _is_package_available("tuna_interpreter")
+def is_import_timer_available():
+    return _is_package_available("import_timer")
 
 
 def is_pynvml_available():