@muellerzr 2024-06-11
@BenjaminBossan what's missing here is the initial tuna check, which is why it's required (see the workflow). 

We can eventually look at gutting it out, sure. However I think having both the visual option for debugging further + the condensed output here is valuable and is why we should still use `tuna` itself.